# hmac: avoid passing the key when calling Reset() on OpenSSL 3.0.3+