--- a/.github/workflows/test.yml
+++ b/.github/workflows/test.yml
@@ -6,7 +6,7 @@ jobs:
       fail-fast: false
       matrix:
         go-version: [1.19.x]
-        openssl-version: [1.0.2, 1.1.0, 1.1.1, 3.0.1]
+        openssl-version: [1.0.2, 1.1.0, 1.1.1, 3.0.1, 3.0.9]
     runs-on: ubuntu-20.04
     steps:
     - name: Install build tools