--- a/hmac.go
+++ b/hmac.go
@@ -115,9 +115,10 @@ func newHMAC3(key []byte, h hash.Hash, md C.GO_EVP_MD_PTR) *opensslHMAC {
 	}
 	var hkey []byte
 	if vMinor == 0 && vPatch <= 2 {
-		// EVP_MAC_init only reset the ctx internal state if a key is passed
-		// when using OpenSSL 3.0.0, 3.0.1, and 3.0.2. New OpenSSL versions
-		// do not have this issue.
+		// EVP_MAC_init only resets the ctx internal state if a key is passed
+		// when using OpenSSL 3.0.0, 3.0.1, and 3.0.2. Save a copy of the key
+		// in the context so Reset can use it later. New OpenSSL versions
+		// do not have this issue so it isn't necessary to save the key.
 		// See https://github.com/openssl/openssl/issues/17811.
 		hkey = make([]byte, len(key))
 		copy(hkey, key)