This PR optimizes `opensslHMAC.Reset` for OpenSSL 3.0.3 and higher versions, in which we don't have to pass the key thanks to https://github.com/openssl/openssl/issues/17811 being fixed.

The perf wins are remarkable:

```cmd
goos: linux
goarch: amd64
pkg: github.com/golang-fips/openssl
cpu: AMD EPYC 7763 64-Core Processor                
                   │   old.txt   │               new.txt               │
                   │   sec/op    │   sec/op     vs base                │
HMACSHA256_32-16     1.338µ ± 0%   1.093µ ± 2%  -18.32% (p=0.000 n=10)
HMACNewWriteSum-16   5.296µ ± 3%   5.374µ ± 4%        ~ (p=0.529 n=10)
geomean              2.661µ        2.423µ        -8.95%

                   │   old.txt    │               new.txt                │
                   │     B/s      │     B/s       vs base                │
HMACSHA256_32-16     22.82Mi ± 0%   27.94Mi ± 2%  +22.44% (p=0.000 n=10)
HMACNewWriteSum-16   5.765Mi ± 3%   5.679Mi ± 4%        ~ (p=0.516 n=10)
geomean              11.47Mi        12.60Mi        +9.83%

                   │  old.txt   │               new.txt               │
                   │    B/op    │    B/op     vs base                 │
HMACSHA256_32-16     80.00 ± 0%   80.00 ± 0%       ~ (p=1.000 n=10) ¹
HMACNewWriteSum-16   352.0 ± 0%   320.0 ± 0%  -9.09% (p=0.000 n=10)
geomean              167.8        160.0       -4.65%
¹ all samples are equal

                   │   old.txt   │               new.txt                │
                   │  allocs/op  │ allocs/op   vs base                  │
HMACSHA256_32-16      3.000 ± 0%   3.000 ± 0%        ~ (p=1.000 n=10) ¹
HMACNewWriteSum-16   10.000 ± 0%   9.000 ± 0%  -10.00% (p=0.000 n=10)
geomean               5.477        5.196        -5.13%
¹ all samples are equal
```